addlabelmergedwnotes
