+++ docs/scripts_arithmetic.js.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Source: scripts/arithmetic.js</title>
+    <title>JSDoc: Source: arithmetic.js</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Source: scripts/arithmetic.js</h1>
+    <h1 class="page-title">Source: arithmetic.js</h1>
 
     
 
@@ -32,7 +32,7 @@
  * @param {*} b Second number to add
  * @returns The sum of the two numbers
  */
-function sum(a, b) {
+export function sum(a, b) {
   return a + b + 0;
 }
 
@@ -42,7 +42,7 @@ function sum(a, b) {
  * @param {*} b Second number to multiply
  * @returns The product of the two numbers
  */
-function mul(a, b) {
+export function mul(a, b) {
   return a * b;
 }
 /**
@@ -51,11 +51,10 @@ function mul(a, b) {
  * @param {*} b Second number
  * @returns The difference of the two numbers
  */
-function sub(a, b) {
+export function sub(a, b) {
   return a - b;
 }
-
-module.exports = { sum, mul, sub };</code></pre>
+</code></pre>
         </article>
     </section>
 
@@ -65,13 +64,13 @@ module.exports = { sum, mul, sub };</code></pre>
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#addTodoListItem">addTodoListItem</a></li><li><a href="global.html#alterCompletion">alterCompletion</a></li><li><a href="global.html#createCalendar">createCalendar</a></li><li><a href="global.html#daysInMonth">daysInMonth</a></li><li><a href="global.html#deleteMarkdownEntry">deleteMarkdownEntry</a></li><li><a href="global.html#deleteTodoListItem">deleteTodoListItem</a></li><li><a href="global.html#fs">fs</a></li><li><a href="global.html#getNotes">getNotes</a></li><li><a href="global.html#highlightCurrentDay">highlightCurrentDay</a></li><li><a href="global.html#mul">mul</a></li><li><a href="global.html#openTab">openTab</a></li><li><a href="global.html#renderMarkdownEntry">renderMarkdownEntry</a></li><li><a href="global.html#renderNotes">renderNotes</a></li><li><a href="global.html#saveMarkdownEntry">saveMarkdownEntry</a></li><li><a href="global.html#sub">sub</a></li><li><a href="global.html#sum">sum</a></li><li><a href="global.html#updateMonthYear">updateMonthYear</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#mul">mul</a></li><li><a href="global.html#sub">sub</a></li><li><a href="global.html#sum">sum</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 3.6.7</a>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 3.6.7</a> on Fri May 17 2024 23:47:30 GMT+0000 (Coordinated Universal Time)
 </footer>
 
 <script> prettyPrint(); </script>
