+++ source/calendar/calendar-script.js
@@ -36,21 +36,6 @@ document.addEventListener("DOMContentLoaded", function () {
     "december",
   ];
   const holidays = [
-  const monthClasses = [
-    "january",
-    "february",
-    "march",
-    "april",
-    "may",
-    "june",
-    "july",
-    "august",
-    "september",
-    "october",
-    "november",
-    "december",
-  ];
-  const holidays = [
     new Date(currentYear, 0, 1), // New Year's Day
     new Date(currentYear, 1, 14), // Valentine's Day
     new Date(currentYear, 6, 4), // Independence Day
@@ -94,12 +79,7 @@ document.addEventListener("DOMContentLoaded", function () {
     let firstDay = new Date(year, month, 1).getDay();
     let lastDay = new Date(year, month + 1, 0).getDay();
     let lastDatePrevMonth = new Date(year, month, 0).getDate();
-    // Get some of the info to build dates
-    let firstDay = new Date(year, month, 1).getDay();
-    let lastDay = new Date(year, month + 1, 0).getDay();
-    let lastDatePrevMonth = new Date(year, month, 0).getDate();
 
-    calendarContainer.innerHTML = ""; // Clear previous calendar
     calendarContainer.innerHTML = ""; // Clear previous calendar
 
     // Add the beginning days of the previous month
@@ -172,7 +152,6 @@ document.addEventListener("DOMContentLoaded", function () {
       currentYear++;
     }
     createCalendar(currentMonth, currentYear);
-    updateSearchButtonColor();
   }
   createCalendar(currentMonth, currentYear);
 
@@ -194,13 +173,5 @@ document.addEventListener("DOMContentLoaded", function () {
       }
     }
   }
-  function updateSearchButtonColor() {
-    const searchBarElement = document.querySelector("search-bar-element");
-    if (searchBarElement) {
-      searchBarElement.updateButtonColor(currentMonth);
-    }
-  }
-
-  updateSearchButtonColor();
+  
 });
-
