+++ source/calendar/index.html
@@ -11,23 +11,10 @@
   </head>
   <body>
     <div class="calendar-header">
-<<<<<<< HEAD
-        <button id="prev-month"><i class="arrow left"></i>Prev</button>
-        <h1 id="current-month"></h1>
-        <!-- Added search bar to the top of the calendar view inside calendar header -->
-        <!-- <div class="search-container">
-          <input type="text" id="search-bar" placeholder="Search labels...">
-          <button id="search-button">Search</button>
-        </div> -->
-        <search-bar-element></search-bar-element>
-        <button id="next-month">Next<i class="arrow right"></i></button>
-      </div>
-=======
       <button id="prev-month"><i class="arrow left"></i>Prev</button>
       <h1 id="current-month"></h1>
       <button id="next-month">Next<i class="arrow right"></i></button>
     </div>
->>>>>>> 195d7d4f471d1c50ce1dd10d2f945d2c807d2d23
     <div class="main-container">
       <div class="calendar-container">
         <div id="day-names" class="day-names">
@@ -92,30 +79,5 @@
         <script src="../components/TodoList.js"></script>
       </div>
     </div>
-
-<<<<<<< HEAD
-    <div>
-      <todo-list-element></todo-list-element>
-      <script src="TodoList.js"></script>
-    </div>
-    </div>
-
-    <div class="label-container">
-      <input type="text" id="label-name" placeholder="Label Name">
-      <input type="color" id="label-color" value="#f0f0f0">
-      <button id="add-label">Add Label</button>
-      <div id="labels-container"></div>
-    </div>
-  
-  <!-- <script src="../Notes/notes-todo-script.js"></script> -->
-  <script src="calendar-script.js"></script>
-  <script src="search-bar.js"></script>
-  <script src="label.js"></script>
 </body>
 </html>
-=======
-    <!-- <script src="../Notes/notes-todo-script.js"></script> -->
-    <script src="calendar-script.js"></script>
-  </body>
-</html>
->>>>>>> 195d7d4f471d1c50ce1dd10d2f945d2c807d2d23
