--- source/main.css
@@ -1,3 +1,0 @@
-p {
-  font-size: 16px;
-}
