+++ source/main.js
@@ -1,15 +1,15 @@
-const { app, BrowserWindow } = require("electron");
-function createWindow() {
+const { app, BrowserWindow } = require('electron')
+function createWindow () {
   let win = new BrowserWindow({
     width: 800,
     height: 600,
     webPreferences: {
       nodeIntegration: true, // turn off nodeIntegration
       contextIsolation: false, // turn on contextIsolationt
-    },
-  });
+    }
+  })
 
-  win.loadFile("calendar/index.html");
+  win.loadFile('calendar/index.html')
 }
 
-app.whenReady().then(createWindow);
+app.whenReady().then(createWindow)
