--- source/notes-test.html
@@ -1,25 +1,0 @@
-<!-- This file is a tempoary file used to demonstrate the creation of two separate label components
-	defined in the label.js file.
-	It uses main.js -->
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Label TEST</title>
-
-    <!-- Label Custom Element -->
-    <script src="./components/label.js" type="module"></script>
-
-    <!-- Main Stylesheets & Scripts -->
-    <link rel="stylesheet" href="./main.css" />
-    <script src="./scripts/generate-labels-ex.js" type="module"></script>
-  </head>
-  <body>
-    <main>
-      <h3>Labels...</h3>
-    </main>
-  </body>
-</html>
