+++ source/package.json
@@ -1,19 +1,25 @@
 {
-  "name": "Dev Journal",
-  "version": "1.0.0",
-  "description": "A Developer Journal That tracks what you do in a day",
-  "main": "main.js",
-  "scripts": {
-      "start": "electron .",
-      "package": "electron-packager . DevJournal --platform=darwin --arch=x64"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-      "electron": "^11.5.0"
-  },
   "devDependencies": {
-      "electron-packager": "^15.5.2"
+    "http-server": "^14.1.1",
+    "jest": "^29.7.0",
+    "jest-puppeteer": "^10.0.1",
+    "jsdoc": "^4.0.3",
+    "prettier": "3.2.5",
+    "puppeteer": "^22.10.0"
+  },
+  "scripts": {
+    "test": "node node_modules/jest/bin/jest.js"
+  },
+  "type": "module",
+  "jest": {
+    "preset": "jest-puppeteer",
+    "testEnvironment": "jest-environment-puppeteer",
+    "transform": {},
+    "verbose": true
+  },
+  "jest-puppeteer": {
+    "server": {
+      "command": "http-server . -p 3000"
+    }
   }
 }
